+++ Runtime/RoadNetwork/Util/Voronoi/RnVoronoiDef.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 574abd43c8ecc8546b582dfb541705e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
